+++ lib/markdown/editor_field.dart
@@ -1,9 +1,12 @@
+import 'dart:ui';
+
 import 'package:flutter/material.dart';
 
 class EditorField extends StatelessWidget {
   const EditorField({
     super.key,
     required this.controller,
+    this.focusNode,
     this.scrollController,
     required this.onChanged,
     this.undoController,
@@ -11,6 +14,7 @@ class EditorField extends StatelessWidget {
   });
 
   final TextEditingController controller;
+  final FocusNode? focusNode;
   final ScrollController? scrollController;
   final Function(String) onChanged;
   final UndoHistoryController? undoController;
@@ -20,9 +24,12 @@ class EditorField extends StatelessWidget {
   Widget build(BuildContext context) {
     return TextField(
       controller: controller,
+      focusNode: focusNode,
       scrollController: scrollController,
       maxLines: null,
       keyboardType: TextInputType.multiline,
+      selectionWidthStyle: BoxWidthStyle.tight,
+      selectionHeightStyle: BoxHeightStyle.tight,
       enableSuggestions: true,
       autocorrect: true,
       autofocus: false,
@@ -30,9 +37,10 @@ class EditorField extends StatelessWidget {
       undoController: undoController,
       style: TextStyle(fontSize: fontSize),
       decoration: InputDecoration(
-          border: InputBorder.none,
-          hintText: 'Type something...',
-          hintStyle: TextStyle(color: Theme.of(context).hintColor)),
+        border: InputBorder.none,
+        hintText: 'Type something...',
+        hintStyle: TextStyle(color: Theme.of(context).hintColor),
+      ),
     );
   }
 }