+++ lib/utils/handlers/item_move.dart
@@ -1,18 +1,14 @@
 import 'package:flutter/material.dart';
-// import 'package:path/path.dart' as path;
-// import 'package:provider/provider.dart';
+import 'package:provider/provider.dart';
 
-// import 'package:printnotes/providers/settings_provider.dart';
+import 'package:printnotes/providers/settings_provider.dart';
 import 'package:printnotes/utils/configs/data_path.dart';
 import 'package:printnotes/utils/storage_system.dart';
 import 'package:printnotes/ui/components/dialogs/select_location.dart';
 
 class ItemMoveHandler {
-  static Future<void> showMoveDialog(
-    BuildContext context,
-    List<Uri> itemUris,
-    Function loadItems,
-  ) async {
+  static Future<void> showMoveDialog(BuildContext context, List<Uri> itemUris,
+      {VoidCallback? onComplete}) async {
     final String? baseDir = await DataPath.selectedDirectory;
     if (baseDir == null) return;
 
@@ -33,11 +29,11 @@ class ItemMoveHandler {
         } else {
           await StorageSystem.moveManyItems(itemUris, newLocationUri);
         }
-        // if (context.mounted) {
-        //   final readSettProv = context.read<SettingsProvider>();
-        //   readSettProv.loadItems(context, readSettProv.currentPath);
-        // }
-        loadItems();
+        if (context.mounted) {
+          final readSettProv = context.read<SettingsProvider>();
+          readSettProv.loadItems(context, readSettProv.currentPath);
+        }
+        onComplete?.call();
       }
     }
   }